+++ src/modules/dateTime/directives/dateTimeBase.ts
@@ -3,7 +3,7 @@ import {Subscription} from 'rxjs';
 
 import {DateTimeInputValue} from '../../../interfaces';
 import {DateTimeInputOutputValue} from '../../../misc/types';
-import {DateTimeDirective} from './dateTime/dateTime.directive';
+import {DateTimeSADirective} from './dateTime/dateTime.directive';
 
 /**
  * Base class for date time directives, contains basic shared data
@@ -26,7 +26,7 @@ export class DateTimeBase<TDate = unknown> implements DateTimeInputValue<TDate>,
     /**
      * Instance of date time shared data, like formats and restrictions
      */
-    protected dateTimeData: DateTimeDirective<TDate> = inject(DateTimeDirective<TDate>);
+    protected dateTimeData: DateTimeSADirective<TDate> = inject(DateTimeSADirective<TDate>);
 
     //######################### public properties - implementation of DateTimeInputValue #########################