+++ src/modules/dateTime/directives/dateTimeMaxValidator/dateTimeMaxValidator.directive.ts
@@ -17,12 +17,13 @@ import {DateTimeBase} from '../dateTimeBase';
         <ExistingProvider>
         {
             provide: NG_VALIDATORS,
-            useExisting: forwardRef(() => DateTimeMaxValidatorDirective),
+            useExisting: forwardRef(() => DateTimeMaxValidatorSADirective),
             multi: true
         },
     ],
+    standalone: true,
 })
-export class DateTimeMaxValidatorDirective<TDate = unknown> extends DateTimeBase<TDate>  implements Validator, OnInit
+export class DateTimeMaxValidatorSADirective<TDate = unknown> extends DateTimeBase<TDate>  implements Validator, OnInit
 {
     //######################### private fields #########################