+++ src/modules/dateTimePicker/modules/dateTimePicker.module.ts
@@ -1,22 +1,25 @@
 import {NgModule} from '@angular/core';
 
 import {DateTimePickerComponent} from '../components';
-import {DateTimePickerDirective} from '../directives';
+import {DateTimePickerSADirective} from '../directives';
 
 /**
  * Module for components that are used as date time picker
  */
 @NgModule(
 {
+    imports:
+    [
+        DateTimePickerSADirective,
+    ],
     declarations:
     [
         DateTimePickerComponent,
-        DateTimePickerDirective,
     ],
     exports:
     [
         DateTimePickerComponent,
-        DateTimePickerDirective,
+        DateTimePickerSADirective,
     ],
 })
 export class DateTimePickerModule